Add files via upload
My knights Tour program : it uses backtracking as its algorithem. 
It runs on a 8x8 chess board but works on any size. For bigger sizes the run time will increase a lot.
The function that runs the knight tour is modulaized in knight.h and the main.c file then calls upon the header function